Rename class to ReturnBookControl
